+++ sci_analysis/test/test_cat_statistics.py
@@ -3,9 +3,10 @@ from numpy import nan
 from numpy.random import seed, randint
 from pandas import Series
 
-from analysis import CategoricalStatistics, NoDataError
+from analysis import CategoricalStatistics
+from analysis.exc import NoDataError
 from data import Categorical, NumberOfCategoriesWarning
-from test.test_categorical import TestWarnings
+from .test_categorical import TestWarnings
 
 
 class MyTestCase(TestWarnings):
