+++ sci_analysis/test/test_equalvariance.py
@@ -2,7 +2,8 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from analysis import EqualVariance, MinimumSizeError, NoDataError
+from analysis import EqualVariance
+from analysis.exc import MinimumSizeError, NoDataError
 
 
 class MyTestCase(unittest.TestCase):
