+++ sci_analysis/test/test_kstest.py
@@ -2,7 +2,8 @@ import unittest
 import scipy.stats as st
 import numpy as np
 
-from analysis import KSTest, MinimumSizeError, NoDataError
+from analysis import KSTest
+from analysis.exc import MinimumSizeError, NoDataError
 
 
 class MyTestCase(unittest.TestCase):
