+++ sci_analysis/test/test_linregress.py
@@ -2,7 +2,8 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from analysis import LinearRegression, MinimumSizeError, NoDataError
+from analysis import LinearRegression
+from analysis.exc import MinimumSizeError, NoDataError
 from data import UnequalVectorLengthError
